Move to the components directory
