+++ apps/app/src/features/mastra/client/components/ChatSidebar/ChatSidebar.tsx
