+++ apps/app/src/features/mastra/client/components/Sidebar/AiAssistant.tsx
