+++ apps/app/src/features/mastra/client/components/Sidebar/AiAssistantSubstance.module.scss
