+++ apps/app/src/features/mastra/client/components/Sidebar/DeleteAiAssistantModal.tsx
@@ -2,7 +2,7 @@ import type React from 'react';
 import { useTranslation } from 'next-i18next';
 import { Button, Modal, ModalBody, ModalFooter, ModalHeader } from 'reactstrap';
 
-import type { AiAssistantHasId } from '../../../openai/interfaces/ai-assistant';
+import type { AiAssistantHasId } from '../../../../openai/interfaces/ai-assistant';
 
 export type DeleteAiAssistantModalProps = {
   isShown: boolean;
